azure-app-functions/requirements.txt
azure-functions
fastapi 